+++ python/ray/tune/BUILD
@@ -234,6 +234,14 @@ py_test(
 )
 
 py_test(
+    name = "test_trial_runner_pg",
+    size = "medium",
+    srcs = ["tests/test_trial_runner_pg.py"],
+    deps = [":tune_lib"],
+    tags = ["exclusive"],
+)
+
+py_test(
     name = "test_trial_runner_callbacks",
     size = "small",
     srcs = ["tests/test_trial_runner_callbacks.py"],
